requirements.txt
rembg
pillow
flask